fix: adb locked when restarting
There is a long-standing bug in Flutter for Windows:
starting a child process will make it where the
child process never dies with us. That means adb will
remain in use by the rogue process, blocking us from
using it.

We approach this from a few angles. For normal window
closing, we can use the `AdbProcessKiller` window
listener. For terminations and crashes, we do the
following: listen for signals and kill the process
there (unreliable on Windows), and store locker files
that persist the PID of the last process that held
the adb lock.

This may only be needed for Windows. Will need to test.

See: https://github.com/dart-lang/sdk/issues/49234